Clean up datamodel-tests a bit
