+++ pkg/datamodel-tests/src/__tests__/runner.test.ts
@@ -5,10 +5,7 @@ import { newTestRunner } from '@runner/runner';
 import * as testDataLoader from '@loaders/smoke-test-loader';
 
 import * as datamodels from '@/tests/implementations';
-import { addBooksToNote, commitNote, deleteNotes, explicitlySetVolumeStock } from '@/tests/tests';
-
-// We're running unit tests on a subset of final, full tests (for different datamodels)
-const tests = { addBooksToNote, commitNote, deleteNotes, explicitlySetVolumeStock };
+import tests from './tests';
 
 describe('Datamodel tests', async () => {
 	// We're currently using the `testDataLoader` from unit tests,
