+++ pkg/datamodel-tests/src/tests/tests.ts
@@ -3,6 +3,58 @@ import { expect } from 'vitest';
 
 import { TestFunction, VolumeQuantityTuple } from '@/types';
 
+// Base functionality
+export const deleteNotes: TestFunction = async (db) => {
+	const w = db.warehouse('wh');
+	const note1 = await w.createInNote();
+	const note2 = await w.createInNote();
+
+	let notes = await w.getNotes();
+	expect(notes.length).toEqual(2);
+
+	// Delete second note
+	await note2.delete();
+	notes = await w.getNotes();
+	expect(notes.length).toEqual(1);
+	expect(await w.getNote(note1._id)).toBeTruthy();
+	await expect(w.getNote(note2._id)).rejects.toThrow();
+};
+
+export const explicitlySetVolumeStock: TestFunction = async (db) => {
+	const w = db.warehouse('wh');
+	const note1 = await w.createInNote();
+	await note1.addVolumes('0001112222', 5);
+
+	let noteFromDB = await w.getNote(note1._id);
+	expect(noteFromDB.books['0001112222']).toEqual(5);
+
+	await note1.setVolumeQuantity('0001112222', 2);
+	noteFromDB = await w.getNote(note1._id);
+	expect(noteFromDB.books['0001112222']).toEqual(2);
+};
+
+export const getFullStock: TestFunction = async (db) => {
+	const science = db.warehouse('science');
+	const horses = db.warehouse('horses');
+
+	const n1 = await science.createInNote();
+	await n1.addVolumes('0123456789', 5);
+	await n1.commit();
+
+	const n2 = await horses.createInNote();
+	await n2.addVolumes('0123456789', 2);
+	await n2.commit();
+
+	const scienceStock = await science.getStock();
+	const horsesStock = await horses.getStock();
+	const fullStock = await db.warehouse().getStock();
+
+	expect(scienceStock).toEqual([{ isbn: '0123456789', quantity: 5 }]);
+	expect(horsesStock).toEqual([{ isbn: '0123456789', quantity: 2 }]);
+	expect(fullStock).toEqual([{ isbn: '0123456789', quantity: 7 }]);
+};
+
+// Tests using the extensive test data
 export const addBooksToNote: TestFunction = async (db, getNotesAndWarehouses) => {
 	const { notes, fullStock } = getNotesAndWarehouses(1);
 	const [noteData] = notes;
@@ -56,65 +108,4 @@ export const commitNote: TestFunction = async (db, getNotesAndWarehouses) => {
 	await note2.commit();
 	wStock = await warehouse.getStock();
 	expect(wStock).toEqual(stock2.books);
-};
-
-export const deleteNotes: TestFunction = async (db) => {
-	const w = db.warehouse('wh');
-	const note1 = await w.createInNote();
-	const note2 = await w.createInNote();
-
-	let notes = await w.getNotes();
-	expect(notes.length).toEqual(2);
-
-	// Delete second note
-	await note2.delete();
-	notes = await w.getNotes();
-	expect(notes.length).toEqual(1);
-	expect(await w.getNote(note1._id)).toBeTruthy();
-	await expect(w.getNote(note2._id)).rejects.toThrow();
-};
-
-export const explicitlySetVolumeStock: TestFunction = async (db) => {
-	const w = db.warehouse('wh');
-	const note1 = await w.createInNote();
-	await note1.addVolumes('0001112222', 5);
-
-	let noteFromDB = await w.getNote(note1._id);
-	expect(noteFromDB.books['0001112222']).toEqual(5);
-
-	await note1.setVolumeQuantity('0001112222', 2);
-	noteFromDB = await w.getNote(note1._id);
-	expect(noteFromDB.books['0001112222']).toEqual(2);
-};
-
-export const getFullStock: TestFunction = async (db) => {
-	const science = db.warehouse('science');
-	const horses = db.warehouse('horses');
-
-	const n1 = await science.createInNote();
-	await n1.addVolumes('0123456789', 5);
-	await n1.commit();
-
-	const n2 = await horses.createInNote();
-	await n2.addVolumes('0123456789', 2);
-	await n2.commit();
-
-	const scienceStock = await science.getStock();
-	const horsesStock = await horses.getStock();
-	const fullStock = await db.warehouse().getStock();
-
-	expect(scienceStock).toEqual([{ isbn: '0123456789', quantity: 5 }]);
-	expect(horsesStock).toEqual([{ isbn: '0123456789', quantity: 2 }]);
-	expect(fullStock).toEqual([{ isbn: '0123456789', quantity: 7 }]);
-};
-
-export const uploadDesignDocuments: TestFunction = async (db) => {
-	const w = db.warehouse();
-	await w.createInNote();
-	await w.createInNote();
-
-	const res = await db._pouch.query('docs/count', { reduce: true });
-	const nDocs = res.rows[0].value;
-
-	expect(nDocs).toEqual(2);
 };
